+++ R/lasupdateheader.r
@@ -119,7 +119,7 @@ lasaddextrabytes = function(las, x, name, desc)
 
 #' @export
 #' @rdname lasaddattribute
-lasaddextrabytes_manual = function(las, x, name, desc, type, offset = NULL, scale = NULL, NA_value = NULL)
+lasaddextrabytes_manual = function(las, x, name, desc, type, offset = NULL, scale = NULL, NA_value = NULL, minmax = TRUE)
 {
   stopifnotlas(las)
   assert_is_a_string(name)
@@ -152,7 +152,8 @@ lasaddextrabytes_manual = function(las, x, name, desc, type, offset = NULL, scal
   }
 
   header     <- as.list(las@header)
-  new_header <- rlas::header_add_extrabytes_manual(header, name, desc, type, offset, scale, min(x, na.rm = TRUE), max(x, na.rm = TRUE), NA_value)
+
+  new_header <- rlas::header_add_extrabytes_manual(header, name, desc, type, offset, scale, NULL, NULL, NA_value)
   new_header <- LASheader(new_header)
   las@header <- new_header
   return(las)